+++ src/main/kotlin/com/recipeapp/recipeserver/repository/AuthorRepository.kt
@@ -3,5 +3,5 @@ package com.recipeapp.recipeserver.repository
 import com.recipeapp.recipeserver.model.Author
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AuthorRepository: JpaRepository<Author, Long> {
+interface AuthorRepository: JpaRepository<Author, Int> {
 }
