+++ src/main/kotlin/com/recipeapp/recipeserver/repository/IngredientRepository.kt
@@ -3,5 +3,5 @@ package com.recipeapp.recipeserver.repository
 import com.recipeapp.recipeserver.model.Ingredient
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface IngredientRepository : JpaRepository<Ingredient, Long> {
+interface IngredientRepository : JpaRepository<Ingredient, Int> {
 }
