+++ src/main/kotlin/com/recipeapp/recipeserver/repository/RecipeIngredientRepository.kt
@@ -3,5 +3,5 @@ package com.recipeapp.recipeserver.repository
 import com.recipeapp.recipeserver.model.RecipeIngredient
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface RecipeIngredientRepository : JpaRepository<RecipeIngredient, Long> {
+interface RecipeIngredientRepository : JpaRepository<RecipeIngredient, Int> {
 }
