+++ src/main/kotlin/com/recipeapp/recipeserver/repository/RecipeRepository.kt
@@ -3,6 +3,6 @@ package com.recipeapp.recipeserver.repository
 import com.recipeapp.recipeserver.model.Recipe
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface RecipeRepository : JpaRepository<Recipe, Long> {
-    fun getById(id: Long): Recipe?
+interface RecipeRepository : JpaRepository<Recipe, Int> {
+    fun getById(id: Int): Recipe?
 }
