+++ src/main/kotlin/com/recipeapp/recipeserver/repository/UnitRepository.kt
@@ -3,5 +3,5 @@ package com.recipeapp.recipeserver.repository
 import com.recipeapp.recipeserver.model.MeasurementUnit
 import org.springframework.data.jpa.repository.JpaRepository
 
-internal interface UnitRepository: JpaRepository<MeasurementUnit, Long> {
+internal interface UnitRepository: JpaRepository<MeasurementUnit, Int> {
 }
