+++ src/main/kotlin/com/recipeapp/recipeserver/service/RecipeService.kt
@@ -4,7 +4,7 @@ import com.recipeapp.recipeserver.model.Recipe
 
 interface RecipeService {
     fun getAllRecipes(): Set<Recipe>
-    fun getRecipeById(id: Long): Recipe?
+    fun getRecipeById(id: Int): Recipe?
     fun addRecipe(recipe: Recipe): Recipe
     fun changeRecipe(recipe: Recipe): Recipe
 }
